+++ src/components/definition-list/dl.component.tsx
@@ -92,4 +92,5 @@ const Dl = ({
   );
 };
 
+Dl.displayName = "Dl";
 export default Dl;