+++ src/components/message/message-content/message-content.style.ts
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+const MessageContentStyle = styled.div`
+  padding: 15px 50px 15px 20px;
+  white-space: pre-wrap;
+  flex: 1;
+
+  .carbon-content__title {
+    margin-bottom: 2px;
+  }
+
+  .carbon-content__body {
+    margin-top: 0px;
+  }
+`;
+
+export default MessageContentStyle;